Move i18n endpoints to directory
