+++ src/endpoints/i18n2/mod.rs
@@ -1,0 +1,6 @@
+pub mod raw_i18n;
+pub mod negotiated_i18n;
+pub mod flat_i18n;
+pub mod untranslated_i18n;
+pub mod used_languages;
+pub mod post_i18n;
